lab2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
fake-useragent==2.2.0
frozenlist==1.6.0
greenlet==3.2.1
idna==3.10
multidict==6.4.3
playwright==1.52.0
propcache==0.3.1
psycopg==3.2.7
psycopg2-binary==2.9.10
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
